New TestCardViewController with buttons
ThumbsUp for pass test, thumbsDown for fail test. When a verb is passed you are left with the cancel button to ensure your answer after seeing the correct answer. If you leave a TestCard without answering it, is canceled, not assigned as a pass or as a failure.

The review mode is now integrated with the test (without the previous review sorting). You can always paginated between the answered cards without changing the results.

All the statistic computation take in account the corrections on a passed and then failed verb