New Performance indicator
Failure moving average. Each failed verb computes as 100 and each passed one as 0. We compute a new failureIndex as a moving average of this values. This index follows closely the failures (two verbs with the same number of failures but with different distribution in time, have different failureIndex, but the same failureRatio)